# Lazily load project submodules