--- a/.github/workflows/pythonpackage.yml
+++ b/.github/workflows/pythonpackage.yml
@@ -28,9 +28,7 @@ jobs:
         VIRES_TOKEN: ${{ secrets.VIRES_AUTH }}
       run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install spacepy  # required for MMS qcotrans tests
-          pip install coveralls
+          pip install '.[test]'
           pip install basemap
           pip install 'mth5>=0.4.9'
           python -m pyspedas.projects.mms.tests.setup_tests
@@ -42,9 +40,7 @@ jobs:
         VIRES_TOKEN: ${{ secrets.VIRES_AUTH }}
       run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install spacepy  # required for MMS qcotrans tests
-          pip install coveralls
+          pip install '.[test]'
           pip install basemap
           pip install 'mth5>=0.4.9'
           python -m pyspedas.projects.mms.tests.setup_tests
@@ -56,9 +52,7 @@ jobs:
         VIRES_TOKEN: ${{ secrets.VIRES_AUTH }}
       run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install spacepy  # required for MMS qcotrans tests
-          pip install coveralls
+          pip install '.[test]'
           pip install basemap
           pip install 'mth5>=0.4.9'
           python -m pyspedas.projects.mms.tests.setup_tests