--- a/.github/workflows/quick_tests.yml
+++ b/.github/workflows/quick_tests.yml
@@ -24,26 +24,17 @@ jobs:
     - name: Install dependencies (Linux)
       run: |
           python -m pip install --upgrade pip
-          pip install .
-          pip install spacepy  # required for MMS qcotrans tests
-          pip install coveralls
-          pip install 'mth5<=0.4.7'
+          pip install '.[test]'
       if: matrix.os == 'ubuntu-latest'
     - name: Install dependencies (Windows)
       run: |
           python -m pip install --upgrade pip
-          pip install .
-          pip install spacepy  # required for MMS qcotrans tests
-          pip install coveralls
-          pip install 'mth5<=0.4.7'
+          pip install '.[test]'
       if: matrix.os == 'windows-latest'
     - name: Install dependencies (macOS)
       run: |
           python -m pip install --upgrade pip
-          pip install .
-          pip install spacepy  # required for MMS qcotrans tests
-          pip install coveralls
-          pip install 'mth5<=0.4.7'
+          pip install '.[test]'
       if: matrix.os == 'macos-latest'
     - name: Lint with flake8
       run: |