--- a/.github/workflows/quick_tests.yml
+++ b/.github/workflows/quick_tests.yml
@@ -24,7 +24,7 @@ jobs:
     - name: Install dependencies (Linux)
       run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install .
           pip install spacepy  # required for MMS qcotrans tests
           pip install coveralls
           pip install 'mth5<=0.4.7'
@@ -32,7 +32,7 @@ jobs:
     - name: Install dependencies (Windows)
       run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install .
           pip install spacepy  # required for MMS qcotrans tests
           pip install coveralls
           pip install 'mth5<=0.4.7'
@@ -40,7 +40,7 @@ jobs:
     - name: Install dependencies (macOS)
       run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install .
           pip install spacepy  # required for MMS qcotrans tests
           pip install coveralls
           pip install 'mth5<=0.4.7'