--- a/.gitignore
+++ b/.gitignore
@@ -206,6 +206,8 @@ venv/
 ENV/
 env.bak/
 venv.bak/
+.pixi
+pixi.lock
 
 # Spyder project settings
 .spyderproject