--- a/pyproject.toml
+++ b/pyproject.toml
@@ -110,6 +110,3 @@ include-package-data = true
 "pyspedas.projects.mms.feeps.sun" = ["*.csv"]
 "pyspedas.projects.mms.mec" = ["*.png"]
 "pyspedas.projects.elfin.epd" = ["el*_epde_cal_data.txt"]
-
-[tool.setuptools.include]
-files = ["*.md", "LICENSE.txt"]