`pyspedas.mission` and `pyspedas.project.mission` still work but now lazily loaded
`import pyspedas.mission` is not supported however.

This pull has breaking changes so we maybe merge it for v2.0

#1050