@Beforerr 2024-11-01
Similar things could be done for `analysis` modules if desired