@Beforerr 2024-11-01
Benchmark shows that first loading speed improve about 0.5s and average 0.15s

❯ hyperfine "python -c 'import pyspedas'"

``` Previous
Benchmark 1: python -c 'import pyspedas'
  Time (mean ± σ):      1.541 s ±  0.236 s    [User: 2.535 s, System: 0.287 s]
  Range (min … max):    1.421 s …  2.205 s    10 runs
 
  Warning: The first benchmarking run for this command was significantly slower than the rest (2.205 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.
```

``` Now
Benchmark 1: python -c 'import pyspedas'
  Time (mean ± σ):      1.401 s ±  0.152 s    [User: 2.520 s, System: 0.266 s]
  Range (min … max):    1.301 s …  1.788 s    10 runs
 
  Warning: The first benchmarking run for this command was significantly slower than the rest (1.788 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.
 
 ```