@Beforerr 2024-11-01
![image](https://github.com/user-attachments/assets/135f6498-4f71-4308-9b24-fe914f93cab9)

Now the mainly loading cost comes from pytplot compareed to previous https://github.com/spedas/pyspedas/issues/1050#issuecomment-2451645905