@Beforerr 2025-08-15
Good to hear the improved performance! 

Besides performance another reason that I would ask to not close this PR is that we can avoid some dependency problem when installing pyspedas (#967). For example, `vires` introduces dependency `tables` and makes `hdf5` a must. This is quite annoying to me for a while (I do not use `vires` so why do I have to bother the problem there by installing a hdf5).

Somehow IDE autocompletion is not a problem for me (using VSCode). Also since we never put `mypy` in the dev deps or check type in CI. Maybe we can just not generate and maintain stub files, or using "if TYPE_CHECKING" clauses for simplicity.