@jameswilburlewis 2025-08-15
Yes, mth5 and projects/secs also have some issues with heavy or hard-to-install external dependencies.  That's something I might be able to address in pyspedas 2.0.0, separately from the lazy loading issue.  I'll open another issue for that.