@Beforerr 2025-08-15
A comprise could be putting packages that introduce significant dep stacks into lazy loading though